Add progress bar to tutorial
