+++ client/www/components/ui.tsx
@@ -1100,31 +1100,3 @@ export function FullscreenLoading() {
     <div className="animate-slow-pulse flex w-full flex-1 flex-col bg-gray-300"></div>
   );
 }
-
-export function PageProgressBar() {
-  const [progress, setProgress] = useState(0);
-
-  useEffect(() => {
-    const updateProgress = () => {
-      const scrollHeight =
-        document.documentElement.scrollHeight - window.innerHeight;
-      const scrolled = window.scrollY;
-      const progress = Math.min((scrolled / scrollHeight) * 100, 100);
-      setProgress(progress);
-    };
-
-    window.addEventListener('scroll', updateProgress);
-    updateProgress();
-
-    return () => window.removeEventListener('scroll', updateProgress);
-  }, []);
-
-  return (
-    <div className="fixed top-0 left-0 right-0 z-50 h-1 bg-gray-200">
-      <div
-        className="h-full bg-gradient-to-r from-blue-500 to-purple-600 transition-all duration-150 ease-out"
-        style={{ width: `${progress}%` }}
-      />
-    </div>
-  );
-}
